[proof] More uniformity on iterators, proof types.
The type invariants here are:

   leaf counts are always usize.  This is because counts can validly be
   zero.  Functions which do not function for zero leaf counts should
   assert so.

   leaf indices are always u64.  The index always starts from zero.
   If there is a valid need to represent an invalid index, use
   Option<u64> or use a count of leaves instead.

   Version is a slightly more abstract type; it could reasonably
   be changed to a tuple or some other struct in the future.  For
   now, try to avoid arithmetic on a version by converting it into
   an index (u64) first.

Closes: #1047
Approved by: zamsden

+++ storage/accumulator/src/tests/mod.rs
@@ -163,9 +163,9 @@ proptest! {
         store.put_many(&writes);
 
         let frozen_subtree_hashes =
-            TestAccumulator::get_frozen_subtree_hashes(&store, leaves.len() as u64).unwrap();
+            TestAccumulator::get_frozen_subtree_hashes(&store, leaves.len()).unwrap();
         let in_mem_acc =
-            InMemoryAccumulator::new(frozen_subtree_hashes, leaves.len() as u64).unwrap();
+            InMemoryAccumulator::new(frozen_subtree_hashes, leaves.len()).unwrap();
         prop_assert_eq!(root_hash, in_mem_acc.root_hash());
     }
 }
